+++ SdlSecurity/sdl_security/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.smartdevicelink.security">
+    package="com.smartdevicelink.sdlsecurity">
 
 <uses-permission android:name="android.permission.INTERNET"/>
