+++ SdlSecurity/sdl_security/src/main/java/com/smartdevicelink/sdlsecurity/SdlSecurity.java
@@ -1,4 +1,4 @@
-package com.smartdevicelink.security;
+package com.smartdevicelink.sdlsecurity;
 
 import android.util.Log;
