+++ SdlSecurity/sdl_security/src/main/java/com/smartdevicelink/sdlsecurity/Tools.java
@@ -1,4 +1,4 @@
-package com.smartdevicelink.security;
+package com.smartdevicelink.sdlsecurity;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
